funding_project/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudscraper==1.2.71
colorama==0.4.6
cron_descriptor==2.0.6
Django==5.2.9
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
kombu==5.6.2
numpy==2.4.0
packaging==25.0
pandas==2.3.3
prompt_toolkit==3.0.52
PyJWT==2.10.1
pyparsing==3.3.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sqlparse==0.5.5
tenacity==9.1.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
psycopg2-binary>=2.9
gunicorn>=20.1.0